refactor(security-architecture): modularize encryption engine and implement core security components with unit tests
