+++ src/secure_document_vault/core/exceptions.py
@@ -1,0 +1,4 @@
+class IntegrityErrorException(Exception):
+    """Excepción lanzada cuando hay un fallo general de integridad o autenticación."""
+
+    pass
