+++ src/secure_document_vault/modules/aead/__init__.py
@@ -1,0 +1,3 @@
+from .engine import AEAD_Engine
+
+__all__ = ["AEAD_Engine"]
