+++ src/secure_document_vault/modules/key_wrapping/__init__.py
@@ -1,0 +1,3 @@
+from .wrapper import ECCKeyWrapper
+
+__all__ = ["ECCKeyWrapper"]
