+++ src/secure_document_vault/modules/randomness/__init__.py
@@ -1,0 +1,3 @@
+from .manager import RandomnessManager
+
+__all__ = ["RandomnessManager"]
