+++ src/secure_document_vault/modules/vault_builder/__init__.py
@@ -1,0 +1,3 @@
+from .builder import VaultBuilder
+
+__all__ = ["VaultBuilder"]
